perf(react-devtools): code-split the panel and gate shadow re-renders
lazy-load the panel impl so it and the compiled css stay out of the production/initial bundle, and only re-render the shadow subtree when its inputs change.
